requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
yfinance>=0.2.40
pydantic>=2.0.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
